# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Upload/GetAllUploads": {
      "post": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_GetAllUploads",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult_UploadDto"
                }
              }
            }
          }
        }
      }
    },
    "/Upload/GetUserUploads": {
      "post": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_GetUserUploads",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoUploadFilterValueTuple"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoUploadFilterValueTuple"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoUploadFilterValueTuple"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoUploadFilterValueTuple"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult_UploadDto"
                }
              }
            }
          }
        }
      }
    },
    "/Upload/GetLibraryUploads": {
      "post": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_GetLibraryUploads",
        "parameters": [
          {
            "name": "libraryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult_UploadDto"
                }
              }
            }
          }
        }
      }
    },
    "/Upload/GetUploadById": {
      "get": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_GetUploadById",
        "parameters": [
          {
            "name": "uploadId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDto"
                }
              }
            }
          }
        }
      }
    },
    "/Upload/AddUpload": {
      "post": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_AddUpload",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AddUploadRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUploadRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUploadRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddUploadRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/Upload/UpdateUpload": {
      "put": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_UpdateUpload",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Upload/DeleteUploadById": {
      "delete": {
        "tags": [
          "Upload"
        ],
        "operationId": "Upload_DeleteUploadById",
        "parameters": [
          {
            "name": "uploadId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddUploadRequestDto": {
        "type": "object",
        "properties": {
          "uploadDto": {
            "$ref": "#/components/schemas/UploadDto"
          },
          "lessonDto": {
            "$ref": "#/components/schemas/LessonDto"
          }
        },
        "additionalProperties": false
      },
      "FileMetadataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "fileType": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checkSum": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LessonDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LessonDto": {
        "type": "object",
        "properties": {
          "mongoId": {
            "type": "string",
            "nullable": true
          },
          "uploadId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "lessonDetails": {
            "$ref": "#/components/schemas/LessonDetailsDto"
          },
          "fileMetadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileMetadataDto"
            },
            "nullable": true
          },
          "ownerId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ownerName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PagedResult_UploadDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UploadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ownerId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "libraryId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "UploadFilter": {
        "type": "object",
        "properties": {
          "uploadIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "ownerIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "libraryIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "uploadTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "dateFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dateTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserDtoUploadFilterValueTuple": {
        "type": "object",
        "additionalProperties": false
      },
...
    }
  }
}